+++ sites/public/src/components/browse/map/ListingsSearchCombined.tsx
@@ -2,17 +2,13 @@ import React, { useState, useEffect, useContext, useRef } from "react"
 import { useMap } from "@vis.gl/react-google-maps"
 import { ListingList, pushGtmEvent, AuthContext } from "@bloom-housing/shared-helpers"
 import { t } from "@bloom-housing/ui-components"
-import {
-  ListingSearchParams,
-  generateSearchQuery,
-  parseSearchString,
-} from "../../../lib/listings/search"
 import { UserStatus } from "../../../lib/constants"
 import styles from "./ListingsSearch.module.scss"
 import { ListingsCombined } from "./ListingsCombined"
 import { MapMarkerData } from "./ListingsMap"
 import { searchListings, searchMapMarkers } from "../../../lib/hooks"
 import {
+  EnumListingFilterParamsComparison,
   ListingFilterParams,
   ListingViews,
 } from "@bloom-housing/shared-helpers/src/types/backend-swagger"
@@ -27,6 +23,7 @@ import {
 import { ListingsMapContext } from "./ListingsMapContext"
 import { useListingsSearchConfigContext } from "./ListingsSearchConfigContext"
 import { useRouter } from "next/router"
+import { ListingQueryBuilder } from "../../../lib/listings/listing-query-builder"
 
 /**
  * This combines the search form with the listings map/list. Listings are updated
@@ -52,20 +49,6 @@ function ListingsSearchCombined() {
 
   const jurisdictionIds = props.jurisdictions.map((jurisdiction) => jurisdiction.id)
   const filterQuery = getFilterQueryFromURL(router.query)
-
-  const [searchFilter] = useState(
-    // TODO: this doesn't work, is it needed?
-    parseSearchString(props.searchString, {
-      bedrooms: null,
-      bathrooms: null,
-      minRent: "",
-      monthlyRent: "",
-      propertyName: "",
-      jurisdictions: props.jurisdictions.map((juris) => juris.name),
-      availability: null,
-      ids: undefined,
-    })
-  )
 
   const [searchResults, setSearchResults] = useState({
     listings: [],
@@ -148,14 +131,12 @@ function ListingsSearchCombined() {
       return
     }
 
-    const modifiedParams: ListingSearchParams = {
-      ...searchFilter,
-      ids: visibleMarkers?.map((marker) => marker.id),
-    }
-
-    // Search the listings by both the filter & the visible markers - but search the markers by only the filter, so that you can scroll out of the currently searched view and still see the markers
-    const listingIdsOnlyQb = generateSearchQuery(modifiedParams)
-    const genericQb = generateSearchQuery(searchFilter)
+    // All searches should only look for "active" listings
+    const genericQb = new ListingQueryBuilder().addFilter(
+      "status",
+      EnumListingFilterParamsComparison["="],
+      "active"
+    )
 
     let newListings = null
     let newMeta
@@ -177,13 +158,21 @@ function ListingsSearchCombined() {
     ) {
       setIsLoading(true)
       const result = await searchListings(
-        isDesktop ? listingIdsOnlyQb : genericQb,
+        genericQb,
         pageSize,
         page,
         listingsService,
         ListingViews.map,
         jurisdictionIds,
-        drawerFilters
+        isDesktop
+          ? [
+              ...drawerFilters,
+              {
+                $comparison: EnumListingFilterParamsComparison.IN,
+                ids: visibleMarkers?.map((marker) => marker.id),
+              },
+            ]
+          : drawerFilters
       )
       hasCompletedInitialListingsCallRef.current = true
       newListings = result.items
@@ -255,38 +244,24 @@ function ListingsSearchCombined() {
 
   useEffect(() => {
     const filterData = decodeQueryToFilterData(router.query)
+    const backendFilters = encodeFilterDataToBackendFilters(filterData).filter(
+      (filter) => filter.name !== ""
+    )
     setFilterState(filterData)
-    setFilterCount(encodeFilterDataToBackendFilters(filterData).length)
-    setIsLoading(false)
+    setDrawerFilters(backendFilters)
+    setFilterCount(backendFilters.length)
   }, [router.query])
 
-  // Re-search when the filter is updated
+  // Re-search when filters change or when entering/leaving mobile
+  // Combined into one effect so the initial load only triggers a single search
   useEffect(() => {
-    async function searchListings() {
+    if (isDesktop === undefined) return
+    async function doSearch() {
       await search(1, true)
     }
-    void searchListings()
+    void doSearch()
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [searchFilter])
-
-  // Re-search when entering mobile
-  useEffect(() => {
-    async function searchListings() {
-      await search(1, true)
-    }
-
-    void searchListings()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isDesktop])
-
-  useEffect(() => {
-    async function searchListings() {
-      await search(1, true)
-    }
-
-    void searchListings()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [drawerFilters])
+  }, [isDesktop, drawerFilters])
 
   // Re-set the view when entering the map on mobile to fit bounds
   useEffect(() => {
@@ -340,15 +315,8 @@ function ListingsSearchCombined() {
   }
 
   const onDrawerSubmit = (data: FilterData) => {
-    const backendFilters = encodeFilterDataToBackendFilters(data).filter(
-      (filter) => filter.name !== ""
-    )
-    setFilterState(data)
-    setDrawerFilters(backendFilters)
-    setFilterCount(backendFilters.length)
     setIsFilterDrawerOpen(false)
     const updatedFilterQuery = encodeFilterDataToQuery(data)
-    setIsFilterDrawerOpen(false)
     if (updatedFilterQuery !== filterQuery) {
       setIsLoading(true)
       void router.push(`/listings?${updatedFilterQuery}`)
@@ -357,9 +325,6 @@ function ListingsSearchCombined() {
 
   const onDrawerClear = (resetFilters: (data: FilterData) => void) => {
     resetFilters({ name: "", monthlyRent: { minRent: "", maxRent: "" } })
-    setFilterState({})
-    setDrawerFilters([])
-    setFilterCount(0)
     setIsFilterDrawerOpen(false)
     void router.push(`/listings`)
   }
@@ -376,7 +341,6 @@ function ListingsSearchCombined() {
     multiselectData: props.multiselectData,
     regions: props.regions,
     listingFeaturesConfiguration: props.listingFeaturesConfiguration,
-    searchFilter,
     searchResults,
     listView,
     setListView,
